cmdtime_d_m_yr_cls
I tweaked it to have the date post as day month year format and also type out cls at the cmd prompt.
I couldn't figure out how to change the position of the blinking prompt to show up after the cls so I deleted that section now it just waits for the minute to end and the screen refreshes. If I get free time Ill attempt to learn actual code and will fix that and add local weather temp if possible. Sadly I have no coding experience or training but know enough to mod existing code so that is what I did.